feature/token-selector-sorting-and-filtering: Added InactiveTokensList component
